@@ -1,6 +1,6 @@
 module github.com/sandia-minimega/minimega/v2
 
-go 1.17
+go 1.20
 
 require (
 	github.com/Harvey-OS/ninep v0.0.0-00010101000000-000000000000
@@ -19,15 +19,15 @@ require (
 	github.com/thoj/go-ircevent v0.0.0-00010101000000-000000000000
 	github.com/twmb/murmur3 v1.1.6
 	github.com/ziutek/telnet v0.0.0-20150427115447-49d9be70897f
-	golang.org/x/crypto v0.21.0
-	golang.org/x/net v0.23.0
-	golang.org/x/sys v0.18.0
+	golang.org/x/crypto v0.33.0
+	golang.org/x/net v0.35.0
+	golang.org/x/sys v0.30.0
 	golang.org/x/tools v0.1.12
 )
 
 require (
 	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
-	golang.org/x/term v0.18.0 // indirect
+	golang.org/x/term v0.29.0 // indirect
 )
 
 replace github.com/Harvey-OS/ninep => github.com/jcrussell/ninep v0.0.0-20180619175724-35ad2879c0a3
